simplify router code
