+++ lib/simple_controller/router/route.rb
@@ -2,18 +2,15 @@ module SimpleController
   class Router
     class Route
       attr_reader :controller_path, :action_name
-      attr_accessor :controller_path_block
 
       def initialize(controller_path, action_name)
         @controller_path, @action_name = controller_path, action_name
       end
 
-      def controller(controller_path_block=nil)
-        controller_path_block ? controller_path_block.call(controller_path) : "#{controller_path}_controller".classify.constantize
-      end
-
-      def call(params, controller_path_block=nil)
-        controller(controller_path_block).call action_name, params
+      def call(*args)
+        controller_path_block = args.last
+        controller_class = controller_path_block ? controller_path_block.call(controller_path) : "#{controller_path}_controller".classify.constantize
+        controller_class.call action_name, *args[0..-2]
       end
     end
   end
